wisherr/requirements.txt
fastapi[all]
sqlalchemy
aiosqlite
jinja2
python-multipart
passlib[bcrypt]
python-dotenv